bots/requirements.txt
python-dotenv~=0.19.2
requests~=2.25.1
twitchio~=2.1.4